// src/model/toir/order/order.ts
interface Order {
    priority: string;
    requestObject: string;
    requestNumber: string;
    dateLimit: string;
    date: string;
    dateCreated: string;
    requestStatus: string;
    vmsRequestNumber: string;
    responsiblePerson: string;
    sourceSystem: string;
    branchOffice: string;
}

export default Order;